5.20.6-优化代码。整理并删除无用消息。插件接口添加发送消息类型。优化服务器主动断开连接时客户端处理。调整C#调用接口。添加并优化日志处理。优化客户端事件调用时机。

+++ TCPClient4.0/main.cpp
@@ -29,7 +29,7 @@ int main()
 		s->WriteInt64(4);
 		s->WriteFloat(5.6f);
 		s->WriteDouble(7.8);
-		s->WriteArray("client", strlen("client"));
+		s->WriteArray("client", (int)strlen("client"));
 		int b[] = { 1, 2, 3, 4, 5 };
 		s->WriteArray(b, sizeof(b));
 		s->Finish(MSG_BYTESTREAM);
